What is wrong?

+++ include/task2.h
@@ -5,7 +5,7 @@
 #ifndef EX_05_TASK1_H
 #define EX_05_TASK1_H
 
-template<class T> T averageArr(T* arr, unsigned int size) {
+template<class T> T averageArr(T* arr, size_t size) {
     T sum = 0;
     for (int i = 0; i < size; i++) {
         sum += arr[i];
